requirements.txt

fastapi
uvicorn[standard]
pydantic>=2.0
SQLAlchemy>=2.0
asyncpg
python-dotenv
alembic     # Para migrar la base de datos 